refactor: move authenticationStore to api client
