+++ packages/api-client/src/helpers/getRequestFromAuthentication.test.ts
