+++ packages/api-client/src/helpers/index.ts
@@ -2,6 +2,7 @@ export * from './concatenateUrlAndPath'
 export * from './createPlaceholderRequest'
 export * from './encodeStringAsBase64'
 export * from './findVariables'
+export * from './getRequestFromAuthentication'
 export * from './humanDiff'
 export * from './mapFromArray'
 export * from './normalizePath'
