+++ packages/api-client/src/stores/index.ts
@@ -1,2 +1,3 @@
-export { useApiClientStore } from './useApiClientStore'
-export { useRequestStore } from './useRequestStore'
+export * from './useApiClientStore'
+export * from './useAuthenticationStore'
+export * from './useRequestStore'
