+++ packages/api-client/src/stores/useAuthenticationStore.ts
@@ -1,6 +1,30 @@
+import type { OpenAPIV3, OpenAPIV3_1 } from '@scalar/openapi-parser'
 import { reactive } from 'vue'
 
-import type { AuthenticationState } from '../types'
+export type AuthenticationState = {
+  preferredSecurityScheme: string | null
+  securitySchemes?:
+    | OpenAPIV3.ComponentsObject['securitySchemes']
+    | OpenAPIV3_1.ComponentsObject['securitySchemes']
+  http: {
+    basic: {
+      username: string
+      password: string
+    }
+    bearer: {
+      token: string
+    }
+  }
+  apiKey: {
+    token: string
+  }
+  oAuth2: {
+    clientId: string
+    scopes: string[]
+    accessToken: string
+    state: string
+  }
+}
 
 export const createEmptyAuthenticationState = (): AuthenticationState => ({
   preferredSecurityScheme: null,
