+++ packages/api-reference/src/components/Content/Authentication/Authentication.vue
@@ -1,9 +1,9 @@
 <script setup lang="ts">
+import { useAuthenticationStore } from '@scalar/api-client'
 import type { OpenAPIV3_1 } from '@scalar/openapi-parser'
 import { computed, watch } from 'vue'
 
 import { hasSecuritySchemes } from '../../../helpers'
-import { useAuthenticationStore } from '../../../stores'
 import { type Spec } from '../../../types'
 import { Card, CardContent, CardHeader } from '../../Card'
 import SecurityScheme from './SecurityScheme.vue'
