+++ packages/api-reference/src/helpers/index.ts
@@ -2,7 +2,6 @@ export * from './deepMerge'
 export * from './getApiClientRequest'
 export * from './getHeadingsFromMarkdown'
 export * from './getLowestHeadingLevel'
-export * from './getRequestFromAuthentication'
 export * from './getUrlFromServerState'
 export * from './getVariableNames'
 export * from './hasModels'
