+++ packages/api-reference/src/stores/index.ts
@@ -1,3 +1,2 @@
-export * from './useAuthenticationStore'
 export * from './useHttpClientStore'
 export * from './useServerStore'
